cmd/sparse-retrieval-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.1
fastembed>=0.7.0
qdrant-client>=1.7.1
numpy>=1.24.0
pydantic>=2.0.0